refactor: extract JdbcDistPollerDao and JdbcInterfaceToNodeCache to daemon-common
Shared JDBC infrastructure for all Spring Boot daemon containers.
Generalized @Scheduled property name from opennms.trapd.* to opennms.daemon.*.
Added AbstractInterfaceToNodeCache.setInstance() for ConvertToEvent static
singleton lookup (Syslogd).
