+++ core/daemon-common/src/test/java/org/opennms/core/daemon/common/JdbcInterfaceToNodeCacheTest.java
@@ -19,7 +19,7 @@
  * language governing permissions and limitations under the
  * License.
  */
-package org.opennms.netmgt.trapd.boot;
+package org.opennms.core.daemon.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
